# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareerConnect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      parameters:
        - in: header
          name: authorization
          schema:
            type: string
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: level
          schema:
            type: string
        - in: query
          name: companyId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/featured:
    get:
      operationId: listFeaturedJobs
      tags: [jobs]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Featured jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: industry
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/top:
    get:
      operationId: listTopCompanies
      tags: [companies]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Top companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      parameters:
        - in: query
          name: companyId
          schema:
            type: integer
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [jobseeker, recruiter, admin]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Job:
      type: object
      required: [id, title, companyId, location, type, level, description]
      properties:
        id:
          type: integer
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        companyLogo:
          type: ["string", "null"]
        location:
          type: string
        type:
          type: string
        level:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        featured:
          type: boolean
        active:
          type: boolean
        postedById:
          type: ["integer", "null"]
        createdAt:
          type: string
        applicationCount:
          type: ["integer", "null"]
    JobInput:
      type: object
      required: [title, companyId, location, type, level, description]
      properties:
        title:
          type: string
        companyId:
          type: integer
        location:
          type: string
        type:
          type: string
        level:
          type: string
        description:
          type: string
        requirements:
          type: string
        salary:
          type: string
        featured:
          type: boolean
        postedById:
          type: integer
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        type:
          type: string
        level:
          type: string
        description:
          type: string
        requirements:
          type: string
        salary:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Company:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        founded:
          type: ["integer", "null"]
        openPositions:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        logo:
          type: string
        description:
          type: string
        website:
          type: string
        industry:
          type: string
        size:
          type: string
        location:
          type: string
        founded:
          type: integer
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string
        description:
          type: string
        website:
          type: string
        industry:
          type: string
        size:
          type: string
        location:
          type: string
        founded:
          type: integer
...
    Review:
      type: object
      required: [id, companyId, userId, rating]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        title:
          type: ["string", "null"]
        pros:
          type: ["string", "null"]
        cons:
          type: ["string", "null"]
        recommend:
          type: boolean
        reviewerName:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [companyId, userId, rating]
      properties:
        companyId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        title:
          type: string
        pros:
          type: string
        cons:
          type: string
        recommend:
          type: boolean
...