# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CareerConnect API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  founded?: number | null;
  /** @nullable */
  openPositions?: number | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt?: string;
}